+++ cmd/web/helpers.go
@@ -6,6 +6,8 @@ import (
 	"net/http"
 	"runtime/debug"
 	"time"
+
+	"github.com/justinas/nosurf"
 )
 
 // The serverError helper writes an error message and stack trace to the errorLog,
@@ -41,9 +43,14 @@ func (app *application) addDefaultData(td *templateData, r *http.Request) *templ
         td = &templateData{}
     }
 
+    // Add the CSRF token to the templateData struct.
+    td.CSRFToken = nosurf.Token(r)
+
     td.CurrentYear = time.Now().Year()
     td.Flash = app.session.PopString(r, "flash")
 
+    // Add the authentication status to the template data.
+    td.IsAuthenticated = app.isAuthenticated(r)
     return td
 }
 
@@ -75,4 +82,14 @@ func (app *application) render(w http.ResponseWriter, r *http.Request, name stri
     // is another time where we pass our http.ResponseWriter to a function that
     // takes an io.Writer.
     buf.WriteTo(w)
+}
+
+// Return true if the current request is from authenticated user, otherwise return false.
+func (app *application) isAuthenticated(r *http.Request) bool {
+    isAuthenticated, ok := r.Context().Value(contextKeyIsAuthenticated).(bool)
+    if !ok {
+        return false
+    }
+
+    return isAuthenticated
 }