+++ cmd/web/main.go
@@ -16,6 +16,10 @@ import (
 	"github.com/golangcollege/sessions"
 )
 
+type contextKey string
+
+const contextKeyIsAuthenticated = contextKey("isAuthenticated")
+
 // Define an application struct to hold the application-wide dependencies for the
 // web application. For now we'll only include fields for the two custom loggers, but
 // we'll add more to it as the build progresses.
@@ -25,6 +29,7 @@ type application struct {
     session *sessions.Session
     snippets *mysql.SnippetModel
     templateCache map[string]*template.Template
+    users *mysql.UserModel
 }
 
 func main() {
@@ -93,6 +98,7 @@ func main() {
         session: session,
         snippets: &mysql.SnippetModel{DB: db},
         templateCache: templateCache,
+        users: &mysql.UserModel{DB: db},
     }
 
     // Initialize a tls.Config struct to hold the non-default TLS settings we want