+++ cmd/web/templates.go
@@ -12,9 +12,11 @@ import (
 // Define a templateData type to act as the holding structure for
 // any dynamic data that we want to pass to our HTML templates.
 type templateData struct {
+    CSRFToken string
     CurrentYear int
     Flash string
     Form *forms.Form
+    IsAuthenticated bool
     Snippet *models.Snippet
     Snippets []*models.Snippet
 }